backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
psycopg[binary]==3.2.9
pydantic==2.11.7
pydantic-settings==2.10.1
email-validator==2.2.0
python-multipart==0.0.20
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.5.0
pandas==2.2.3
openpyxl==3.1.5
boto3==1.39.17
httpx==0.28.1
pytest==8.4.1
google-api-python-client==2.159.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
